Backend/requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
Django==5.2.9
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyJWT==2.12.1
redis==7.4.0
sqlparse==0.5.4
tzdata==2025.2
